Removal of play.Project in order to simplify project declarations.
