+++ framework/src/sbt-plugin/src/main/scala/play/Project.scala
@@ -3,75 +3,12 @@
  */
 package play
 
-import sbt.{ Project => _, _ }
+import sbt._
 import sbt.Keys._
 
 import com.typesafe.sbt.SbtNativePackager.packageArchetype
-import com.typesafe.sbt.web.SbtWeb
-import com.typesafe.sbt.jse.{SbtJsEngine, SbtJsTask}
+import com.typesafe.sbt.jse.SbtJsTask
 import com.typesafe.sbt.webdriver.SbtWebDriver
-
-object Project extends PlayExceptions with play.Keys with PlayReloader with PlayCommands
-    with PlayRun with play.Settings with PlayPositionMapper with PlaySourceGenerators {
-
-  private lazy val sbtWebSettings =
-    SbtWeb.projectSettings ++
-      SbtJsEngine.projectSettings ++
-      SbtJsTask.projectSettings ++
-      SbtWebDriver.projectSettings
-
-  private lazy val commonSettings: Seq[Setting[_]] =
-    packageArchetype.java_server ++
-      defaultSettings ++
-      intellijCommandSettings ++
-      Seq(testListeners += testListener) ++
-      Seq(
-        scalacOptions ++= Seq("-deprecation", "-unchecked", "-encoding", "utf8"),
-        javacOptions in Compile ++= Seq("-encoding", "utf8", "-g")
-      )
-
-  private[play] lazy val javaSettings: Seq[Setting[_]] =
-    commonSettings ++
-      eclipseCommandSettings(JAVA) ++
-      defaultJavaSettings ++
-      Seq(libraryDependencies += javaCore)
-
-  @deprecated("Use the PlayJava plugin instead i.e. project...addPlugins(PlayJava).", "2.3")
-  lazy val playJavaSettings: Seq[Setting[_]] = sbtWebSettings ++ javaSettings
-
-  private[play] lazy val scalaSettings: Seq[Setting[_]] =
-    commonSettings ++
-      eclipseCommandSettings(SCALA) ++
-      defaultScalaSettings
-
-  @deprecated("Use the PlayScala plugin instead i.e. project...addPlugins(PlayScala).", "2.3")
-  lazy val playScalaSettings: Seq[Setting[_]] = sbtWebSettings ++ scalaSettings
-
-  @deprecated("Use the PlayJava or PlayScala plugins instead e.g. project...addPlugins(PlayJava).", "2.3")
-  def apply(name: String, applicationVersion: String = "1.0", dependencies: Seq[ModuleID] = Nil, path: File = file("."), settings: => Seq[Setting[_]] = Seq()): sbt.Project = {
-    lazy val playSettings = if (dependencies.contains(javaCore)) playJavaSettings else playScalaSettings
-
-    lazy val projectSettings: Seq[Setting[_]] = Seq(
-      version := applicationVersion,
-      libraryDependencies ++= dependencies
-    )
-
-    sbt.Project(name, path)
-      .settings(playSettings: _*)
-      .settings(projectSettings: _*)
-      .settings(settings: _*)
-  }
-}
-
-/**
- * Declares the default imports for Play plugins.
- */
-object PlayImport {
-
-  import play.Keys
-
-  object PlayKeys extends Keys
-}
 
 /**
  * Base plugin for Play projects. Declares common settings for both Java and Scala based Play projects.
@@ -82,11 +19,24 @@ object Play
   with PlayReloader
   with PlayCommands
   with PlayRun
-  with play.Settings
+  with play.PlaySettings
   with PlayPositionMapper
   with PlaySourceGenerators {
 
   override def requires = SbtJsTask && SbtWebDriver
+
+  val autoImport = play.PlayImport
+
+
+  override def projectSettings =
+    packageArchetype.java_server ++
+      defaultSettings ++
+      intellijCommandSettings ++
+      Seq(testListeners += testListener) ++
+      Seq(
+        scalacOptions ++= Seq("-deprecation", "-unchecked", "-encoding", "utf8"),
+        javacOptions in Compile ++= Seq("-encoding", "utf8", "-g")
+      )
 }
 
 /**
@@ -99,9 +49,12 @@ object Play
 object PlayJava extends AutoPlugin {
   override def requires = Play
 
-  override def projectSettings = Project.javaSettings
-
-  val autoImport = PlayImport
+  import Play._
+  import Play.autoImport._
+  override def projectSettings =
+    eclipseCommandSettings(JAVA) ++
+    defaultJavaSettings ++
+    Seq(libraryDependencies += javaCore)
 }
 
 /**
@@ -114,7 +67,8 @@ object PlayJava extends AutoPlugin {
 object PlayScala extends AutoPlugin {
   override def requires = Play
 
-  override def projectSettings = Project.scalaSettings
-
-  val autoImport = PlayImport
+  import Play._
+  override def projectSettings =
+    eclipseCommandSettings(SCALA) ++
+    defaultScalaSettings
 }
