+++ framework/src/sbt-plugin/src/sbt-test/play-sbt-plugin/secret/project/Build.scala
@@ -3,20 +3,16 @@
  */
 import sbt._
 import Keys._
-import com.typesafe.sbt.packager.Keys._
-import play.Project._
 
 object ApplicationBuild extends Build {
 
   val appName = "secret-sample"
   val appVersion = "1.0-SNAPSHOT"
 
-  val appDependencies = Seq(
-  )
-
   val Secret = """(?s).*application.secret="(.*)".*""".r
 
-  val main = play.Project(appName, appVersion, appDependencies).settings(
+  val main = Project(appName, file(".")).addPlugins(play.PlayScala).settings(
+    version := appVersion,
     TaskKey[Unit]("check-secret") := {
       val file = IO.read(baseDirectory.value / "conf/application.conf")
       file match {
