+++ framework/test/integrationtest-java/project/Build.scala
@@ -3,7 +3,8 @@
  */
 import sbt._
 import Keys._
-import play.Project._
+import play.PlayJava
+import play.Play.autoImport._
 
 object ApplicationBuild extends Build {
 
@@ -17,9 +18,9 @@ object ApplicationBuild extends Build {
       "org.hamcrest" % "hamcrest-all" % "1.3"
     )
 
-    val main = play.Project(appName, appVersion, appDependencies).settings(
-      // Add your own project settings here
-      compile in Test <<= PostCompile(Test)
+    val main = Project(appName, file(".")).addPlugins(PlayJava).settings(
+      version := appVersion, 
+      libraryDependencies ++= appDependencies
     )
 
 }
