+++ samples/java/websocket-chat/build.sbt
@@ -1,9 +1,7 @@
-import play.Project._
-
 name := "websocket-chat"
 
 version := "1.0"
 
 javacOptions += "-Xlint:deprecation"     
 
-playJavaSettings
+lazy val root = (project in file(".")).addPlugins(PlayJava)
