+++ samples/java/zentasks/build.sbt
@@ -1,9 +1,7 @@
-import play.Project._
-
 name := "zentask"
 
 version := "1.0"
 
 libraryDependencies ++= Seq(javaJdbc, javaEbean)     
 
-playJavaSettings
+lazy val root = (project in file(".")).addPlugins(PlayJava)
